Thursday. Got the project to open and added some more objects, including another plant enemy and the goal flag, though it isn't functional yet.

+++ Assets/EnemyDamage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EnemyDamage : MonoBehaviour
 {
-    [SerializeField] public float damage;
+    [SerializeField] public int damage;
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
